Health bar values are now universal
Health bar now scales based off of a percentage of HP rather than raw HP.

+++ Assets/Scripts/HealthBar.cs
@@ -10,7 +10,7 @@ public class HealthBar : MonoBehaviour
     public Slider healthBar; // Health bar slider
 
     // Method to set the value of the health bar
-    public void SetHP(int hp)
+    public void SetHP(float hp)
     {
         healthBar.value = hp;
     }
